ai-verifier/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
google-generativeai>=0.5.4
slowapi>=0.1.9
httpx>=0.27.0
Pillow>=10.3.0
